+++ database/migrations/2024_06_07_063105_create_purchases_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->uuid('id')->primary();
             $table->string('title');
             $table->string('type');
-            $table->integer('value');
+            $table->decimal('value', 10, 2);
             $table->text('description');
             $table->dateTime('date');
             $table->foreignUuid('user_id')->constrained()->onDelete('cascade');
